fix:other users can see thier own expenses

+++ src/main/java/com/start/expense_tracker/service/impl/ExpenseServiceImpl.java
@@ -12,8 +12,10 @@ import com.start.expense_tracker.dto.ExpenseResponse;
 import com.start.expense_tracker.dto.ExpenseSummaryResponse;
 import com.start.expense_tracker.entity.Category;
 import com.start.expense_tracker.entity.Expense;
+import com.start.expense_tracker.entity.User;
 import com.start.expense_tracker.mapper.ExpenseMapper;
 import com.start.expense_tracker.repository.ExpensesRepository;
+import com.start.expense_tracker.security.utils.SecurityUtils;
 import com.start.expense_tracker.service.ExpenseService;
 
 import lombok.RequiredArgsConstructor;
@@ -22,13 +24,15 @@ import lombok.RequiredArgsConstructor;
 public class ExpenseServiceImpl implements ExpenseService {
 	
 	private final ExpensesRepository expenseRepository;
-	
+	private final SecurityUtils securityUtils; 
 
 
 
 	@Override
 	public ExpenseResponse createExpense(ExpenseRequest request) {
+		  User currentUser = securityUtils.getCurrentUser();
 			Expense expense = ExpenseMapper.toEntity(request);
+			expense.setUser(currentUser);
 			Expense saved = expenseRepository.save(expense);
 			return ExpenseMapper.toResponse(saved);
 		
@@ -36,7 +40,14 @@ public class ExpenseServiceImpl implements ExpenseService {
 
 	@Override  
 	public ExpenseResponse updateExpense(Long id,ExpenseRequest request) {
+		 User currentUser = securityUtils.getCurrentUser();
 		Expense expense = expenseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Expense not found with id: " + id));
+		
+		if(!expense.getUser().getId().equals(currentUser.getId()) ) {
+			throw new RuntimeException(
+	                "You are not authorized to update this expense!");
+		}
+		
 		expense.setDescription(request.getDescription());
 		expense.setAmount(request.getAmount());
 		expense.setDate(request.getDate());
@@ -46,27 +57,38 @@ public class ExpenseServiceImpl implements ExpenseService {
 
 	@Override
 	public void deleteExpense(Long id) {
-
+		User currentUser = securityUtils.getCurrentUser();
 	    Expense expense = expenseRepository.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("Expense not found with id: " + id));
+	    if (!expense.getUser().getId().equals(currentUser.getId())) {
+            throw new RuntimeException(
+                "You are not authorized to delete this expense!");
+        }
 
 	    expenseRepository.delete(expense);
 	}
 
 	@Override
 	public List<ExpenseResponse> getAllExpenses() {
-		 List<Expense> expenses = expenseRepository.findAll();
-		 return expenses.stream().map(ExpenseMapper::toResponse).toList();
+		User currentUser = securityUtils.getCurrentUser();
+		 return expenseRepository.findByUser(currentUser)
+	                .stream()
+	                .map(ExpenseMapper::toResponse)
+	                .toList();
 	}
 
 	@Override
 	public ExpenseSummaryResponse getTotalSummary() {
-		List<Expense> expenses = expenseRepository.findAll();
+		User currentUser = securityUtils.getCurrentUser();
+		List<Expense> expenses =
+	            expenseRepository.findByUser(currentUser);
+		
 		BigDecimal totalAmount = expenses.stream()
 				.map(Expense::getAmount)
 				.filter(a -> a != null)
 				.reduce(BigDecimal.ZERO, BigDecimal::add);
 		long totalCount = expenses.size();
+		
 		return ExpenseSummaryResponse.builder()
 				.totalAmount(totalAmount)
 				.totalCount(totalCount)
@@ -75,12 +97,17 @@ public class ExpenseServiceImpl implements ExpenseService {
 
 	@Override
 	public ExpenseSummaryResponse getMonthlySummary(YearMonth yearMonth) {
-		List<Expense> byDateBetween = expenseRepository.findByDateBetween(yearMonth.atDay(1), yearMonth.atEndOfMonth());
+		User currentUser = securityUtils.getCurrentUser();
+		
+		List<Expense> byDateBetween = expenseRepository.findByUserAndDateBetween(currentUser,yearMonth.atDay(1), yearMonth.atEndOfMonth());
+		
 		BigDecimal totalAmount = byDateBetween.stream()
 				.map(Expense::getAmount)
 				.filter(a -> a != null)
 				.reduce(BigDecimal.ZERO, BigDecimal::add);
+		
 		long totalCount = byDateBetween.size();
+		
 		return ExpenseSummaryResponse.builder()
 				.totalAmount(totalAmount)
 				.totalCount(totalCount)
@@ -89,8 +116,13 @@ public class ExpenseServiceImpl implements ExpenseService {
 
 	@Override
 	public List<ExpenseResponse> getByCategory(Category category) {
-		List<Expense> byCategory = expenseRepository.findByCategory(category);
-		return byCategory.stream().map(ExpenseMapper::toResponse).toList();
+		User currentUser = securityUtils.getCurrentUser();
+		
+		return expenseRepository
+                .findByUserAndCategory(currentUser, category)
+                .stream()
+                .map(ExpenseMapper::toResponse)
+                .toList();
 	}
 
 	@Override
@@ -98,11 +130,18 @@ public class ExpenseServiceImpl implements ExpenseService {
 		if (ids == null || ids.isEmpty()) {
 			throw new ResourceNotFoundException("No ids provided");
 		}
-		List<Expense> byIds = expenseRepository.findByIdIn(ids);
-		if(byIds.isEmpty()) {
-			throw new ResourceNotFoundException("No expenses found with the provided ids");
-		}
-		return byIds.stream().map(ExpenseMapper::toResponse).toList();
+		 User currentUser = securityUtils.getCurrentUser();
+		 
+		 List<Expense> expenses =
+		            expenseRepository.findByUserAndIdIn(currentUser, ids);
+
+		 if (expenses.isEmpty()) {
+	            throw new ResourceNotFoundException(
+	                "No expenses found with the provided ids");
+	        }
+		 return expenses.stream()
+	                .map(ExpenseMapper::toResponse)
+	                .toList();
 	}
